different r/ask... commands are in separate classes
